+++ .playground/playgroud.tsx
@@ -1,50 +1,75 @@
 import React, { Fragment } from 'react';
-import { Axis, Chart, getAxisId, getSpecId, Position, ScaleType, Settings, BarSeries } from '../src';
+import {
+  Axis,
+  Chart,
+  getAxisId,
+  getSpecId,
+  Position,
+  ScaleType,
+  Settings,
+  BarSeries,
+  LineSeries,
+  AreaSeries,
+} from '../src';
 
 export class Playground extends React.Component {
   render() {
     return (
       <Fragment>
         <div className="chart">
-          <Chart>
+          <Chart className="story-chart">
             <Settings
-              showLegend={true}
               theme={{
-                axes: {
-                  gridLineStyle: {
-                    horizontal: {
-                      stroke: 'red',
-                      strokeWidth: 0.5,
-                      opacity: 1,
-                      dash: [0, 0],
-                    },
-                    vertical: {
-                      stroke: 'blue',
-                      strokeWidth: 0.5,
-                      opacity: 1,
-                      dash: [4, 4],
-                    },
+                areaSeriesStyle: {
+                  point: {
+                    visible: true,
                   },
                 },
               }}
+              xDomain={{
+                max: 3.8,
+              }}
             />
             <Axis
-              id={getAxisId('y')}
+              id={getAxisId('bottom')}
+              position={Position.Bottom}
+              title={'Bottom axis'}
+              showOverlappingTicks={true}
+            />
+            <Axis
+              id={getAxisId('left')}
+              title={'Left axis'}
               position={Position.Left}
               domain={{
-                min: 50,
-                max: 250,
+                max: 5,
               }}
-              showGridLines
             />
-            <Axis showGridLines id={getAxisId('x')} position={Position.Bottom} />
+
             <BarSeries
               id={getSpecId('bar')}
-              yScaleType={ScaleType.Linear}
               xScaleType={ScaleType.Linear}
+              yScaleType={ScaleType.Linear}
               xAccessor={0}
               yAccessors={[1]}
-              data={[[0, 100], [1, 50], [3, 400], [4, 250], [5, 235]]}
+              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
+            />
+
+            <LineSeries
+              id={getSpecId('line')}
+              xScaleType={ScaleType.Linear}
+              yScaleType={ScaleType.Linear}
+              xAccessor={0}
+              yAccessors={[1]}
+              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
+            />
+
+            <AreaSeries
+              id={getSpecId('area')}
+              xScaleType={ScaleType.Linear}
+              yScaleType={ScaleType.Linear}
+              xAccessor={0}
+              yAccessors={[1]}
+              data={[[0, 1], [1, 2], [2, 10], [3, 4], [4, 5]]}
             />
           </Chart>
         </div>