+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [13.1.1](https://github.com/elastic/elastic-charts/compare/v13.1.0...v13.1.1) (2019-09-28)
+
+
+### Bug Fixes
+
+* **rendering:** out-of-domain rendering of points/bars/lines/areas ([#395](https://github.com/elastic/elastic-charts/issues/395)) ([b6fee52](https://github.com/elastic/elastic-charts/commit/b6fee52)), closes [#386](https://github.com/elastic/elastic-charts/issues/386)
+
+# [13.1.0](https://github.com/elastic/elastic-charts/compare/v13.0.1...v13.1.0) (2019-09-27)
+
+
+### Features
+
+* **axis:** add option for integer only axis ticks ([#389](https://github.com/elastic/elastic-charts/issues/389)) ([4fcfe3c](https://github.com/elastic/elastic-charts/commit/4fcfe3c)), closes [#387](https://github.com/elastic/elastic-charts/issues/387)
+
+## [13.0.1](https://github.com/elastic/elastic-charts/compare/v13.0.0...v13.0.1) (2019-09-27)
+
+
+### Bug Fixes
+
+* x-scale for linear band charts ([#384](https://github.com/elastic/elastic-charts/issues/384)) ([daa3b55](https://github.com/elastic/elastic-charts/commit/daa3b55))
+
 # [13.0.0](https://github.com/elastic/elastic-charts/compare/v12.1.0...v13.0.0) (2019-09-19)
 
 
@@ -8,7 +29,7 @@
 
 ### BREAKING CHANGES
 
-* **axis:** `theme.axes.tickLineStyle.visible` is now required (default base is false)
+* **axis:** `theme.axes.tickLineStyle.visible` is now required (default base is `true`)
 
 # [12.1.0](https://github.com/elastic/elastic-charts/compare/v12.0.2...v12.1.0) (2019-09-19)
 
@@ -41,7 +62,7 @@
 
 ### BREAKING CHANGES
 
-* **theme:** Added `GridLineStyle` to `Theme` (`theme.gridLineStyle.horizontal` and `theme.gridLineStyle.vertical`)
+* **theme:** Added `GridLineStyle` to `Theme` (`theme.axes.gridLineStyle.horizontal` and `theme.axes.gridLineStyle.vertical`)
 
 * add gridLineStyle to AxisConfig
 * add chartTheme vs axisSpec